backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
pypdf
python-docx
python-dotenv
supabase
